guix-zephyr: Bug fix: packages no longer need to specify outputs
This commit adds an explicit check for the debug output before the
build gexp is generated. It seems `guix build` passes '("out") as an
explicit output list in the case no such list is provided by the
package.

I personally think this is a bug in the behavior of Guix. The build
function provides default values for outputs and `guix build` should
not presume what these should be.